Import and use gotham
